🎨 Uplift challengeState to App
